segnn/requirement.txt
torch-scatter==1.3.1
torch-sparse==0.4.0
torch-cluster==1.4.4
torch-spline-conv==1.1.0
torch-geometric==1.3.2
tqdm==4.31.1
tensorboardX==1.4







